Let LocalDataStore Option Detecting Local File Change Automatically  (#245)
**Context**
This requirement is mentioned by atlassian in comments. They want us
"constantly watching the local file and any changes in it should be
automatically pulled in and updated. This helps in changing the FF value
locally and testing the feature and writing test for it."

Thread Chat Link: https://atl-statsig.atlassian.net/browse/SIG-157

**Code Change**
This diff:

1. Refactor the previous `localDataStore` I wrote to fit this use case
better.
3. Add fileWatch to detect whether users made any changes to the local
file, if so, updating the `specStore`.

Left more details on comments