fix: change lockableArray into ConcurrentQueue (#404)
**Context:** 
We have a team experiencing blocking in their service when calling check
gate due to exposures sent to the lockable array. They are experiencing
this when setting the id as a random UUID which we have told them to
change to a finite number of ids which they will be testing, but we
believe this locking is not necessary where you could instead use a
ConcurrentLinkedQueue.