feat: sampling (#406)
### Context: 

This pr added sampling to those non-analytic events. (we sampled on
gates/config/layer exposures no matter it's a manual exposure event or
not)

There are some conditions you need to meet to decide whether we should
sample or not. Details are inside method `getSamplingDecisionAndDetails`
and `shouldSkipSampling`, read that carefully


![image](https://github.com/user-attachments/assets/ab979fd4-e255-4814-964c-10c177f2b99b)